Fixed signed/unsigned conversion in txt2las.c
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40864

+++ apps/txt2las.c
@@ -479,7 +479,7 @@ int main(int argc, char *argv[])
     /* create output file name if none specified and no piped output requested */
     if (file_name_out == NULL && file_name_in != NULL)
     {
-        int len = strlen(file_name_in);
+        int len = (int)strlen(file_name_in);
         file_name_out = strdup(file_name_in);
 
         while (len > 0 && file_name_out[len] != '.')
